feat: replace supply category dialog with side panel
