+++ src/features/supplyCategories/SettingsButton.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from 'react';
 import { IconButton } from '@/src/components/IconButton/iconButton';
-import { SupplyCategorySettingsDialog } from './SupplyCategorySettingsDialog';
+import SupplyCategoryPanel from './SupplyCategoryPanel';
 
 export function SettingsButton() {
     const [open, setOpen] = useState(false);
@@ -10,7 +10,7 @@ export function SettingsButton() {
     return (
         <>
             <IconButton icon={<i className="fa-solid fa-cog"></i>} title="Settings" onClick={() => setOpen(true)} />
-            {open && <SupplyCategorySettingsDialog open={open} onClose={() => setOpen(false)} />}
+            {open && <SupplyCategoryPanel open={open} onClose={() => setOpen(false)} />}
         </>
     );
 }
